+++ src/heap/mod.rs
@@ -7,6 +7,8 @@
 
 mod array_heap;
 mod binary_heap;
+mod pair_queue;
 
 pub use array_heap::{ArrayHeap, ArrayHeapConfig, Pair};
 pub use binary_heap::MinHeap;
+pub use pair_queue::{Entry, PairQueue};